sistema-distribuido/etl/requirements.txt
Flask
psycopg2-binary
cx_Oracle